+++ CHANGELOG.md
@@ -108,6 +108,7 @@ Released with 1.0.0-beta.37 code base.
 - ``BN`` and ``BigNumber`` objects are now supported by the ``abi.encodeParameter(s)`` method (#3238) 
 - ``getPendingTransactions`` added to web3-eth package (#3239)
 - Revert instruction handling added which can get activated with the ``handleRevert`` module property (#3248)
+- The ``receipt`` does now exist as property on the error object for transaction related errors (#3259)
 
 ### Changed